% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getAffinityMatrix}
\alias{getAffinityMatrix}
\title{Get Affinity Matrix}
\usage{
getAffinityMatrix(distance2, knn, nc)
}
\arguments{
\item{distance2}{The distance matrix for each observation}

\item{knn}{The number of nearest neighbors}

\item{nc}{The number of observations for data in this class}
}
\value{
an affinity matrix - the larger the element in the matrix, the closer two data points are
}
\description{
This function returns an affinity matrix within knn-nearest neighbors from the distance matrix.
}
