% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getMetricOfType}
\alias{getMetricOfType}
\title{Get Requested Type of Transforming Metric}
\usage{
getMetricOfType(metric, eigVec, eigVal, total)
}
\arguments{
\item{metric}{The type of metric to be requested}

\item{eigVec}{The eigenvectors of the problem}

\item{eigVal}{The eigenvalues of the problem}

\item{total}{The number of total rows to be used for weighting denominator}
}
\value{
The transformation metric in requested type
}
\description{
This function returns the requested type of transforming metric.
}
