% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metric_transformed_data_viz_3d.R
\name{MetricTransformedViz}
\alias{MetricTransformedViz}
\title{3D Visualization on Metric Transformed Data}
\usage{
MetricTransformedViz(dat, method = c("lfda", "klfda"))
}
\arguments{
\item{dat}{The original data.}

\item{method}{The metric learning method used to transform the original data. It can be "lfda" or "klfda".}
}
\description{
This function performs metric learning on the given dataset and runs
a 3D visualization on the transformed data.
}
\seealso{
See \code{\link{lfda}} and \code{\link{klfda}} for the metric learning method used for this visualization.
}

