% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lfda_visualization.R
\name{plot.lfda}
\alias{plot.lfda}
\title{3D Visualization for LFDA/KLFDA Result}
\usage{
\method{plot}{lfda}(x, labels, cleanText = FALSE, ...)
}
\arguments{
\item{x}{The lfda/klfda result.}

\item{labels}{A list of class labels used for lfda/klfda training.}

\item{cleanText}{A boolean value to specify whether to make the labels in the plot cleaner (default: FALSE)}

\item{...}{Additional arguments}
}
\description{
This function plot 3 dimensions of the lfda/klfda result.
}
\seealso{
See \code{\link{lfda}} and \code{\link{klfda}} for the metric learning method used for this visualization.
}

