% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lfda.R
\name{print.lfda}
\alias{print.lfda}
\title{Print an lfda object}
\usage{
\method{print}{lfda}(x, ...)
}
\arguments{
\item{x}{The result from lfda function, which contains a transformed data and a transforming}

\item{...}{ignored}
}
\description{
Print an lfda object
}

